api/pyproject.toml
[project]
name = "trading-dashboard-api"
version = "0.1.0"
description = "FastAPI backend for trading dashboard"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.0",
    "aiosqlite>=0.21.0",
    "websockets>=14.0",
    "siliconmetatrader5>=1.1.0",
    "numpy>=2.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pyright>=1.1.390",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"

[tool.pytest.ini_options]
asyncio_mode = "auto"
